feat(GET /objects): compile filter grammar once
